feat-component: add chart line componentc
